requirements.txt
langchain>=0.3.1
langchain-community>=0.3.1
chromadb>=0.4.22
gradio>=4.19.2
pypdf>=4.0.1
python-dotenv>=1.0.0
sentence-transformers>=2.5.1
pydantic>=2.6.1 
pydantic-settings>=2.0.0
openai>=1.91.0
tiktoken>=0.9.0
langchain-chroma>=0.1.2
langchain-openai>=0.3.25
beautifulsoup4>=4.12.0
requests>=2.31.0